Extract visitor to utils
